Moved np.ndarray->dpt.usm_ndarray cast-and-copy out to separate file
